moved launchValidation to Utils
