+++ src/oauthHandlers/tokenHandlerStrategyClasses/saveDocumentStrategies/saveDocumentLaunchStrategy.js
@@ -1,4 +1,4 @@
-const { hashString } = require("../../../utils");
+const { hashString, launchValidation } = require("../../../utils");
 const { getUnixTime } = require("date-fns");
 
 class SaveDocumentLaunchStrategy {
@@ -13,7 +13,7 @@ class SaveDocumentLaunchStrategy {
       launch = document.launch;
       if (
         !tokens.scope.split(" ").includes("launch/patient") &&
-        !this.launchValidation(launch)
+        !launchValidation(launch)
       ) {
         throw {
           status: 400,
@@ -42,30 +42,6 @@ class SaveDocumentLaunchStrategy {
         error
       );
     }
-  }
-  launchValidation(launch) {
-    if (launch === null || launch === "") return false;
-    var base64regex = /^([0-9a-zA-Z+/]{4})*(([0-9a-zA-Z+/]{2}==)|([0-9a-zA-Z+/]{3}=))?$/;
-    if (base64regex.test(launch)) {
-      let decodedLaunch = JSON.parse(
-        Buffer.from(launch, "base64").toString("ascii")
-      );
-      if (
-        !decodedLaunch["patient"] ||
-        typeof decodedLaunch["patient"] != typeof "string"
-      ) {
-        return false;
-      }
-    } else {
-      if (launch === "" || Object.keys(launch).length === 0) {
-        return false;
-      } else if (
-        !launch.hasOwnProperty("patient") ||
-        typeof launch["patient"] != typeof "string"
-      )
-        return false;
-    }
-    return true;
   }
 }
