+++ src/oauthHandlers/tokenHandlerStrategyClasses/tokenHandlerClient.js
@@ -2,6 +2,7 @@ const {
   rethrowIfRuntimeError,
   hashString,
   minimalError,
+  launchValidation,
 } = require("../../utils");
 const { translateTokenSet } = require("../tokenResponse");
 const { staticRefreshTokenIssueCounter } = require("../../metrics");
@@ -114,11 +115,7 @@ class TokenHandlerClient {
       }
 
       if (document.launch) {
-        if (
-          this.saveDocumentToDynamoStrategy.launchValidation(
-            document.launch
-          ) === false
-        ) {
+        if (launchValidation(document.launch) === false) {
           return {
             statusCode: 400,
             responseBody: {
